Move extr filters to attrib_util.py
This helps with a backport to the v3 branch
